responsiveness done

+++ src/Pages/Dashboard/MyOrders/MyOrders.js
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from 'react';
 import useAuth from '../../../hooks/useAuth';
 import SingleOrder from '../SingleOrder/SingleOrder';
 import './MyOrders.css'
+import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline';
 
 const MyOrders = () => {
     const {user} = useAuth();
@@ -16,30 +17,34 @@ const MyOrders = () => {
     },[])
 
     const handleDelete = (id) =>{
-        fetch(`https://radiant-citadel-36252.herokuapp.com/orders/${id}`, {
-            method: 'DELETE'
-        })
-        .then(res => res.json())
-        .then(data => {
-            if(data.deletedCount > 0){
-            const remainingOrders = myOrders.filter(order => order._id!== id);
-            setMyOrders(remainingOrders);
-            }
-        } )
+        const proceed = window.confirm('Are you sure you want to delete?')
+        if(proceed){
+            fetch(`https://radiant-citadel-36252.herokuapp.com/orders/${id}`, {
+                method: 'DELETE'
+            })
+            .then(res => res.json())
+            .then(data => {
+                if(data.deletedCount > 0){
+                const remainingOrders = myOrders.filter(order => order._id!== id);
+                setMyOrders(remainingOrders);
+                }
+            } )
+        }
+       
     
     }
     
     return (
         <div className="container p-3">
-            <h1>My Orders</h1>
-            <h4>Total Orders : {length}</h4>
-             <div className='orders-display p-2 rounded'>
+            <h1 className='fw-bold text-center mt-5 mb-3'>My Orders</h1>
+            <h4 className='mb-3'>Total Orders : {length}</h4>
+             <div className='orders-display p-3 rounded'>
              {
                  myOrders.map(order => <SingleOrder
                  key={order._id}
                  order = {order}
                 >
-                <button className="btn button" onClick={() => handleDelete(order._id)}>Cancel Order</button>
+                <button className="btn button-action" onClick={() => handleDelete(order._id)}><DeleteOutlineIcon/> Cancel Order</button>
                  </SingleOrder>)
              }
              </div>
